_import now works, shellchecked clean, del, extract, help & import

+++ src/_del.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 function _delete() {
-    selected_parcels=$(ls "$PARCELS_FOLDER" | fzf --multi --preview 'echo "$PARCELS_FOLDER/{}"')
+    selected_parcels=$(ls "${PARCELS_FOLDER:?Parcels folder not set.}" | fzf --multi --preview 'echo "$PARCELS_FOLDER/{}"')
     for parcel in $selected_parcels; do
         if [[ ! -f "$parcel" ]]; then
             
@@ -15,7 +15,7 @@ function _delete() {
 
             out "Deleting parcel: $parcel"
 
-            rm -r "$PARCELS_FOLDER/${parcel}"
+            rm -r "${PARCELS_FOLDER:?Parcels folder not set}/${parcel}"
         fi
     done 
 }
